Auxiliary function, gives mask for matrix for directed, undirected etc.

Auxiliary function, gives mask for matrix for directed, undirected etc.

mat2vec.ix(mat, directed, selfloops)

Arguments

mat
matrix

directed
a boolean argument specifying whether object is directed or not.

selfloops
a boolean argument specifying whether the model should incorporate selfloops.

Value

a boolean matrix that can be used to mask adjacency matrices.

Examples

data('adj_karate')
mat2vec.ix(adj_karate, FALSE, FALSE)

#>          Mr Hi Actor 2 Actor 3 Actor 4 Actor 5 Actor 6 Actor 7 Actor 8 Actor 9
#> Mr Hi    FALSE    TRUE    TRUE    TRUE    TRUE    TRUE    TRUE    TRUE    TRUE
#> Actor 2  FALSE   FALSE    TRUE    TRUE    TRUE    TRUE    TRUE    TRUE    TRUE
#> Actor 3  FALSE   FALSE   FALSE    TRUE    TRUE    TRUE    TRUE    TRUE    TRUE
#> Actor 4  FALSE   FALSE   FALSE   FALSE    TRUE    TRUE    TRUE    TRUE    TRUE
#> Actor 5  FALSE   FALSE   FALSE   FALSE   FALSE    TRUE    TRUE    TRUE    TRUE
#> Actor 6  FALSE   FALSE   FALSE   FALSE   FALSE   FALSE    TRUE    TRUE    TRUE
#> Actor 7  FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE    TRUE    TRUE
#> Actor 8  FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE    TRUE
#> Actor 9  FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 10 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 11 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 12 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 13 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 14 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 15 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 16 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 17 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 18 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 19 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 20 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 21 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 22 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 23 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 24 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 25 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 26 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 27 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 28 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 29 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 30 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 31 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 32 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> Actor 33 FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#> John A   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE   FALSE
#>          Actor 10 Actor 11 Actor 12 Actor 13 Actor 14 Actor 15 Actor 16
#> Mr Hi        TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 2      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 3      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 4      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 5      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 6      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 7      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 8      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 9      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 10    FALSE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 11    FALSE    FALSE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 12    FALSE    FALSE    FALSE     TRUE     TRUE     TRUE     TRUE
#> Actor 13    FALSE    FALSE    FALSE    FALSE     TRUE     TRUE     TRUE
#> Actor 14    FALSE    FALSE    FALSE    FALSE    FALSE     TRUE     TRUE
#> Actor 15    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE     TRUE
#> Actor 16    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 17    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 18    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 19    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 20    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 21    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 22    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 23    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 24    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 25    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 26    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 27    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 28    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 29    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 30    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 31    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 32    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 33    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> John A      FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#>          Actor 17 Actor 18 Actor 19 Actor 20 Actor 21 Actor 22 Actor 23
#> Mr Hi        TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 2      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 3      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 4      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 5      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 6      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 7      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 8      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 9      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 10     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 11     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 12     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 13     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 14     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 15     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 16     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 17    FALSE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 18    FALSE    FALSE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 19    FALSE    FALSE    FALSE     TRUE     TRUE     TRUE     TRUE
#> Actor 20    FALSE    FALSE    FALSE    FALSE     TRUE     TRUE     TRUE
#> Actor 21    FALSE    FALSE    FALSE    FALSE    FALSE     TRUE     TRUE
#> Actor 22    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE     TRUE
#> Actor 23    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 24    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 25    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 26    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 27    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 28    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 29    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 30    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 31    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 32    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 33    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> John A      FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#>          Actor 24 Actor 25 Actor 26 Actor 27 Actor 28 Actor 29 Actor 30
#> Mr Hi        TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 2      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 3      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 4      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 5      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 6      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 7      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 8      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 9      TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 10     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 11     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 12     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 13     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 14     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 15     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 16     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 17     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 18     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 19     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 20     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 21     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 22     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 23     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 24    FALSE     TRUE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 25    FALSE    FALSE     TRUE     TRUE     TRUE     TRUE     TRUE
#> Actor 26    FALSE    FALSE    FALSE     TRUE     TRUE     TRUE     TRUE
#> Actor 27    FALSE    FALSE    FALSE    FALSE     TRUE     TRUE     TRUE
#> Actor 28    FALSE    FALSE    FALSE    FALSE    FALSE     TRUE     TRUE
#> Actor 29    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE     TRUE
#> Actor 30    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 31    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 32    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> Actor 33    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#> John A      FALSE    FALSE    FALSE    FALSE    FALSE    FALSE    FALSE
#>          Actor 31 Actor 32 Actor 33 John A
#> Mr Hi        TRUE     TRUE     TRUE   TRUE
#> Actor 2      TRUE     TRUE     TRUE   TRUE
#> Actor 3      TRUE     TRUE     TRUE   TRUE
#> Actor 4      TRUE     TRUE     TRUE   TRUE
#> Actor 5      TRUE     TRUE     TRUE   TRUE
#> Actor 6      TRUE     TRUE     TRUE   TRUE
#> Actor 7      TRUE     TRUE     TRUE   TRUE
#> Actor 8      TRUE     TRUE     TRUE   TRUE
#> Actor 9      TRUE     TRUE     TRUE   TRUE
#> Actor 10     TRUE     TRUE     TRUE   TRUE
#> Actor 11     TRUE     TRUE     TRUE   TRUE
#> Actor 12     TRUE     TRUE     TRUE   TRUE
#> Actor 13     TRUE     TRUE     TRUE   TRUE
#> Actor 14     TRUE     TRUE     TRUE   TRUE
#> Actor 15     TRUE     TRUE     TRUE   TRUE
#> Actor 16     TRUE     TRUE     TRUE   TRUE
#> Actor 17     TRUE     TRUE     TRUE   TRUE
#> Actor 18     TRUE     TRUE     TRUE   TRUE
#> Actor 19     TRUE     TRUE     TRUE   TRUE
#> Actor 20     TRUE     TRUE     TRUE   TRUE
#> Actor 21     TRUE     TRUE     TRUE   TRUE
#> Actor 22     TRUE     TRUE     TRUE   TRUE
#> Actor 23     TRUE     TRUE     TRUE   TRUE
#> Actor 24     TRUE     TRUE     TRUE   TRUE
#> Actor 25     TRUE     TRUE     TRUE   TRUE
#> Actor 26     TRUE     TRUE     TRUE   TRUE
#> Actor 27     TRUE     TRUE     TRUE   TRUE
#> Actor 28     TRUE     TRUE     TRUE   TRUE
#> Actor 29     TRUE     TRUE     TRUE   TRUE
#> Actor 30     TRUE     TRUE     TRUE   TRUE
#> Actor 31    FALSE     TRUE     TRUE   TRUE
#> Actor 32    FALSE    FALSE     TRUE   TRUE
#> Actor 33    FALSE    FALSE    FALSE   TRUE
#> John A      FALSE    FALSE    FALSE  FALSE